+++ libraries/container/src/main/java/androidx/media3/container/Mp4Box.java
@@ -227,6 +227,15 @@ public abstract class Mp4Box {
   @SuppressWarnings("ConstantCaseForConstants")
   public static final int TYPE_tfhd = 0x74666864;
 
+  @SuppressWarnings({"ConstantCaseForConstants", "IdentifierName"})
+  public static final int TYPE_tfra = 0x74667261;
+
+  @SuppressWarnings({"ConstantCaseForConstants", "IdentifierName"})
+  public static final int TYPE_mfro = 0x6d66726f;
+
+  @SuppressWarnings({"ConstantCaseForConstants", "IdentifierName"})
+  public static final int TYPE_mfra = 0x6d667261;
+
   @SuppressWarnings("ConstantCaseForConstants")
   public static final int TYPE_trex = 0x74726578;